+++ src/main/java/dev/hoyin1600p/vhaccelerator/backport/BackportOwner.java
@@ -1,0 +1,8 @@
+package dev.hoyin1600p.vhaccelerator.backport;
+
+public enum BackportOwner {
+    VHA,
+    MODERNFIX,
+    DISABLED,
+    UNAVAILABLE
+}