+++ src/main/java/dev/hoyin1600p/vhaccelerator/backport/ModernFixOwnership.java
@@ -1,0 +1,8 @@
+package dev.hoyin1600p.vhaccelerator.backport;
+
+public enum ModernFixOwnership {
+    ABSENT,
+    ACTIVE,
+    INACTIVE,
+    UNKNOWN
+}